feat(resource): add GetAllExchanges feature
