+++ src/FinnHub.MCP.Server/Resources/Exchanges/ExchangesResource.cs
@@ -8,6 +8,8 @@
 using System.Diagnostics.CodeAnalysis;
 using System.Net.Mime;
 using System.Text.Json.Serialization;
+using FinnHub.MCP.Server.Application.Exchanges;
+using FinnHub.MCP.Server.Application.Exchanges.Features.GetAllExchanges;
 using FinnHub.MCP.Server.Application.Models;
 
 namespace FinnHub.MCP.Server.Resources.Exchanges;
@@ -16,7 +18,6 @@ namespace FinnHub.MCP.Server.Resources.Exchanges;
 /// MCP resource that provides stock exchange listings from Finnhub.
 /// Uses <see cref="BaseResource.CreateResponse{T}"/> to serialize JSON data for clients.
 /// </summary>
-[ExcludeFromCodeCoverage]
 public sealed class ExchangesResource : BaseResource
 {
     private static readonly ResourceTemplate s_template = new()
@@ -65,7 +66,7 @@ public sealed class ExchangesResource : BaseResource
         var uri = request.Params.Uri;
 
         // TODO: Replace stub data with real Finnhub API call logic.
-        var responsePayload = new ExchangeResponse
+        var responsePayload = new ExchangesResponse
         {
             Exchanges =
             [
@@ -85,9 +86,9 @@ public sealed class ExchangesResource : BaseResource
         };
 
         var result = CreateResponse(
-            new Result<ExchangeResponse>().Success(responsePayload),
+            new Result<ExchangesResponse>().Success(responsePayload),
             uri,
-            ResourceJsonContext.Default.ResultExchangeResponse);
+            ResourceJsonContext.Default.ResultExchangesResponse);
 
         return ValueTask.FromResult(result);
     }
@@ -99,103 +100,5 @@ public sealed class ExchangesResource : BaseResource
     WriteIndented = true,
     DefaultIgnoreCondition = JsonIgnoreCondition.WhenWritingNull,
     PropertyNameCaseInsensitive = true)]
-[JsonSerializable(typeof(Result<ExchangeResponse>))]
+[JsonSerializable(typeof(Result<ExchangesResponse>))]
 public partial class ResourceJsonContext : JsonSerializerContext;
-
-/// <summary>
-/// Represents a paginated response that contains one or more stock exchange records.
-/// </summary>
-[ExcludeFromCodeCoverage]
-public sealed class ExchangeResponse
-{
-    /// <summary>
-    /// Gets the list of stock exchanges returned by the API.
-    /// </summary>
-    [JsonPropertyName("exchanges")]
-    public IReadOnlyList<Exchange> Exchanges { get; init; } = [];
-
-    /// <summary>
-    /// Gets the number of exchanges in the response.
-    /// </summary>
-    [JsonPropertyName("total_count")]
-    public int TotalCount => this.Exchanges.Count;
-
-    /// <summary>
-    /// Gets whether there are any exchanges returned.
-    /// </summary>
-    [JsonPropertyName("has_results")]
-    public bool HasResults => this.TotalCount > 0;
-}
-
-/// <summary>
-/// Represents detailed information about a single stock exchange.
-/// </summary>
-[ExcludeFromCodeCoverage]
-public sealed class Exchange
-{
-    /// <summary>
-    /// Gets the exchange code or symbol (e.g. "NYSE", "AB").
-    /// </summary>
-    [JsonPropertyName("code")]
-    public required string ExchangeCode { get; init; }
-
-    /// <summary>
-    /// Gets the full human-readable name of the exchange.
-    /// </summary>
-    [JsonPropertyName("name")]
-    public required string ExchangeName { get; init; }
-
-    /// <summary>
-    /// Gets the Market Identifier Code (MIC), per ISO 10383.
-    /// </summary>
-    [JsonPropertyName("mic")]
-    public string? MicCode { get; init; }
-
-    /// <summary>
-    /// Gets the IANA time zone identifier for the exchange's location.
-    /// </summary>
-    [JsonPropertyName("time_zone")]
-    public string? TimeZone { get; init; }
-
-    /// <summary>
-    /// Gets the scheduled start time of the pre‑market session.
-    /// </summary>
-    [JsonPropertyName("pre_market_hours")]
-    public string? PreMarketHours { get; init; }
-
-    /// <summary>
-    /// Gets the regular market trading hours.
-    /// </summary>
-    [JsonPropertyName("trading_hours")]
-    public string? TradingHours { get; init; }
-
-    /// <summary>
-    /// Gets the scheduled hours of the post‑market session.
-    /// </summary>
-    [JsonPropertyName("post_market_hours")]
-    public string? PostMarketHours { get; init; }
-
-    /// <summary>
-    /// Gets the date or time when the exchange is closed (if applicable).
-    /// </summary>
-    [JsonPropertyName("close_date")]
-    public string? CloseDate { get; init; }
-
-    /// <summary>
-    /// Gets the ISO 2‑letter country code where the exchange is located.
-    /// </summary>
-    [JsonPropertyName("country_code")]
-    public required string CountryCode { get; init; }
-
-    /// <summary>
-    /// Gets the full country name where the exchange is located.
-    /// </summary>
-    [JsonPropertyName("country_name")]
-    public required string CountryName { get; init; }
-
-    /// <summary>
-    /// Gets the URL for more information about the exchange.
-    /// </summary>
-    [JsonPropertyName("url")]
-    public required string Url { get; init; }
-}
